Make volatile vertices tracking inside Graph
What I wanted to achieve: special set of nodes (volatile ones), which I add however I want, but they hold the invariant that they are subset of the graph.

Few rejected ideas:
* this.removeNodeFromVolatileInFormula at beginning of each setCell* method (because it's repetitive)
* this.exchangeOrAddNode / this.exchangeNode / ... methods in DependencyGraph, which in addition of using Graph, they also remove the node from volatile set if necessary. Not bad, but I wanted to avoid additional methods in DependencyGraph, as it's big anyway
* Delegator composed of Graph, which tracks the changes of tracked nodes. Not bad either, but one more function call and one more layer to understand. Also I don't know good way of doing delegation for most methods.

That way, Graph has one additional not-the-most-pleasant structure,
specialNodes, but it still only knows that this is "subset of
nodes". I don't care if we'd add more such subsets if needed, I think