Add LOG10 function
The "templateWithOneCoercedToNumberArgument" is taken from
TrigonometryPlugin, as it's very common formula pattern. I think I
will extract that somewhere, most likely to FunctionPlugin. I've
called it template, because it is a template for the very generic
formula behaviour and tells nothing about what actually function does.

Initially I was hesitant about moving such functions out of the
respective plugin, because I thought it will make it harder to read,
as part of the formula behaviour will be in one place, and one in the
other. But it turns out, that I can name the function accurately
enough that the reader will know what it does.

I don't want to make any kind of "generic template", because I don't
know whether such generic abstraction would be useful, and it most
likely hurt readability.