Add licenseKeyValidityState prop to the Config
Adds "licenseKeyValidityState" property to the Config class. The
property holds a state about checking the license key validity. The
property is protected from modification by defining its getter only.

Additionally, the commit adds TS definition for the licenseKey checker
helper and the result of the expired message is formatted better.