Fix duplicated messages about deprecated options usage (#882)
The commit fixes an issue related to the doubled or tripled printed deprecated
warning messages.

After the changes once the Hyperformula is initialized the warn messages are
logged only once. The messages are logged again only if the
`engine.updateConfig()` method is called and only when the deprecated options
are passed.

Do prints deprecated warn messages:
* `HyperFormula.buildEmpty({ binarySearchThreshold: 10 })`
* `engine.updateConfig({ binarySearchThreshold: 10 })`

Do not print deprecated warn messages:
* `engine.updateConfig({ precisionRounding: 10 })`
* `engine.rebuildAndRecalculate()`

I had a problem with tests in the browser. It turned out that there is a
problem with unregistering Jasmine spies that interfere with other tests. To
fix it up the commit upgrades the Jasmine to the latest version.

The change solves the problem with Handsontable integration where its Formulas
plugin calls the HF rebuildAndRecalculate method. The method internally merges
the Config object with default values (e.g. `binarySearchThreshold: 20`) that
once passed in the constructor logs again warning messages about the usage of
deprecated options. See handsontable/handsontable#8771.