Update basic-operations.md (#893)
* Update basic-operations.md

Your documentation says, "A sheet can be removed by using the removeSheet method. To do that you need to pass a mandatory parameter: <b>the name of a sheet to be removed.</b> However, in practice, this method requires a number, not a string. I used ```const sheetid = hfInstance.getSheetId("my_sheet");
  if (typeof sheetid == "number") hfInstance.removeSheet(sheetid);)``` to remove the sheet

* Update basic-operations.md